Started registration
