+++ source/baseapp/appconfig.json
@@ -1,1 +1,5 @@
-{"api-token": "1759223276:AAHs8LJsHvn0rkHNCm9oktDRyf97aGtGjsA", "domain": "0dafcdf5a014.ngrok.io", "debug": true}
+{
+    "api-token": "1759223276:AAHs8LJsHvn0rkHNCm9oktDRyf97aGtGjsA",
+    "domain": "78f97a8a8ccd.ngrok.io",
+    "debug": true
+}
