+++ source/baseapp/bot/handlers/common_objects.py
@@ -1,0 +1,6 @@
+from ..bot_init import bot
+from ..dbmodels.states import *
+from ..dbmodels.validators import *
+from ..dbmodels.chat import Chat
+from ..dbmodels.auth import ScheduleUser
+from telebot import types
