+++ source/baseapp/bot/handlers/login.py
@@ -1,0 +1,44 @@
+from .common_objects import *
+
+@bot.callback_query_handler(lambda call: call.data == 'c_login')
+def button_login(call: types.CallbackQuery):
+    callback_message = call.message
+    print("............................................")
+    try:
+        chat_id = callback_message.chat.id
+        chat = Chat.objects.get(chat_id=chat_id)
+        bot.send_message(chat_id, "Enter username:")
+        chat.state = State.LOGING_IN_UNAME.value
+        chat.save()
+    except Exception as ex:
+        bot.send_message(callback_message.chat.id,
+                         "Something went wrong\nTry again")
+
+@bot.message_handler(content_types=['text'], func=lambda message: onstate(message.chat.id, State.LOGING_IN_UNAME))
+def login_username_input(message):
+    message_input = message.text
+    chat_id = message.chat.id
+    chat = Chat.objects.get(chat_id=chat_id)
+    try:
+        user = ScheduleUser.objects.get(username=message_input)
+        chat.connected_user = user
+        chat.save()
+    except Exception:
+        user = None
+    chat.state = State.LOGING_IN_PASS.value
+    chat.save()
+    bot.send_message(chat_id, "Enter password:")
+
+@bot.message_handler(content_types=['text'], func=lambda message: onstate(message.chat.id, State.LOGING_IN_PASS))
+def login_password_input(message):
+    message_input = message.text
+    chat_id = message.chat.id
+    chat = Chat.objects.get(chat_id=chat_id)
+    if chat.connected_user != None and chat.connected_user.check_password(message_input) == True:
+        chat.authorised = True
+        bot.send_message(chat_id, "Success!")
+    else:
+        chat.connected_user = None
+        bot.send_message(chat_id, "Login failed!\nIvalid username or password")
+    chat.state = State.NO_ACTIONS.value
+    chat.save()
